fix(Acore-update): Uses new functionality from refactors in AC
Updated the scripts to use the new versions of the functions that were refactored.

https://github.com/azerothcore/azerothcore-wotlk/commit/50f8f145d224037be9a29c4390c5088816639868#diff-4ed19160746a0c7a1783c225f72ff686a9d5e654767a5ed6d8f08b1f05d2da18
https://github.com/azerothcore/azerothcore-wotlk/commit/2f7f9bd72f188298d8805e1faaa4019c1e03016c#diff-215889510ed8663cd10c316fed36d4bbaa0a2aab2fbb0b1ad328efec1011d645

+++ src/AdvancementMgr.cpp
@@ -108,9 +108,9 @@ int32 AdvancementMgr::LoadAdvancementRanks() {
         uint32 maxIncrease2 = fields[11].Get<uint32>();
         uint32 minIncrease3 = fields[12].Get<uint32>();
         uint32 maxIncrease3 = fields[13].Get<uint32>();
-        uint32 chanceCost1 = fields[14].Get<uint32>();
-        uint32 chanceCost2 = fields[15].Get<uint32>();
-        uint32 chanceCost3 = fields[16].Get<uint32>();
+        int32 chanceCost1 =  static_cast<int32>(fields[14].Get<uint32>());
+        int32 chanceCost2 = static_cast<int32>(fields[15].Get<uint32>());
+        int32 chanceCost3 =  static_cast<int32>(fields[16].Get<uint32>());
 
         // Should add validator... but let's do it without and trust in the o-DB-Wan-kenobe
         MpAdvancements advancement = static_cast<MpAdvancements>(advancementId);
